Created `chaintracker` package and moved code
