+++ pkg/chaintracker/chainfollower.go
@@ -1,4 +1,4 @@
-package broker
+package chaintracker
 
 import (
 	"context"
